+++ frontend/src/features/whiteboard/views/board/_components/nodes/SpanAnnotationNode.tsx
@@ -58,7 +58,9 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isCodeSpanAnnotationEdge)
       .filter((edge) => edge.target === `spanAnnotation-${props.data.spanAnnotationId}`) // isEdgeForThisSpanAnnotation
       .filter((edge) => parseInt(edge.source.split("-")[1]) !== codeId); // isEdgeForIncorrectCode
-    reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    if (edgesToDelete.length > 0) {
+      reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    }
 
     // checks which code nodes are already in the graph and adds edges to the correct node
     const existingCodeNodeIds = reactFlowInstance
@@ -66,9 +68,11 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isCodeNode)
       .map((code) => code.data.codeId);
     if (existingCodeNodeIds.includes(codeId)) {
-      reactFlowInstance.addEdges([
-        createCodeSpanAnnotationEdge({ codeId, spanAnnotationId: props.data.spanAnnotationId }),
-      ]);
+      const newEdge = createCodeSpanAnnotationEdge({ codeId, spanAnnotationId: props.data.spanAnnotationId });
+      const edgeExists = reactFlowInstance.getEdges().some((edge) => edge.id === newEdge.id);
+      if (!edgeExists) {
+        reactFlowInstance.addEdges([newEdge]);
+      }
     }
   }, [props.data.spanAnnotationId, reactFlowInstance, code.data]);
 
@@ -82,7 +86,9 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isSdocSpanAnnotationEdge)
       .filter((edge) => edge.target === `spanAnnotation-${props.data.spanAnnotationId}`) // isEdgeForThisSpanAnnotation
       .filter((edge) => parseInt(edge.source.split("-")[1]) !== sdocId); // isEdgeForIncorrectSdoc
-    reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    if (edgesToDelete.length > 0) {
+      reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    }
 
     // checks which sdoc nodes are already in the graph and adds edges to the correct node
     const existingSdocNodeIds = reactFlowInstance
@@ -90,9 +96,11 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isSdocNode)
       .map((sdoc) => sdoc.data.sdocId);
     if (existingSdocNodeIds.includes(sdocId)) {
-      reactFlowInstance.addEdges([
-        createSdocSpanAnnotationEdge({ sdocId, spanAnnotationId: props.data.spanAnnotationId }),
-      ]);
+      const newEdge = createSdocSpanAnnotationEdge({ sdocId, spanAnnotationId: props.data.spanAnnotationId });
+      const edgeExists = reactFlowInstance.getEdges().some((edge) => edge.id === newEdge.id);
+      if (!edgeExists) {
+        reactFlowInstance.addEdges([newEdge]);
+      }
     }
   }, [props.data.spanAnnotationId, reactFlowInstance, annotation.data]);
 
@@ -106,7 +114,9 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isMemoSpanAnnotationEdge)
       .filter((edge) => edge.target === `spanAnnotation-${props.data.spanAnnotationId}`) // isEdgeForThisSpanAnnotation
       .filter((edge) => parseInt(edge.source.split("-")[1]) !== memoId); // isEdgeForIncorrectMemo
-    reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    if (edgesToDelete.length > 0) {
+      reactFlowInstance.deleteElements({ edges: edgesToDelete });
+    }
 
     // checks which memo nodes are already in the graph and adds edge to the correct node
     const existingMemoNodeIds = reactFlowInstance
@@ -114,9 +124,11 @@ export function SpanAnnotationNode(props: NodeProps<SpanAnnotationNode>) {
       .filter(isMemoNode)
       .map((memo) => memo.data.memoId);
     if (existingMemoNodeIds.includes(memoId)) {
-      reactFlowInstance.addEdges([
-        createMemoSpanAnnotationEdge({ memoId, spanAnnotationId: props.data.spanAnnotationId }),
-      ]);
+      const newEdge = createMemoSpanAnnotationEdge({ memoId, spanAnnotationId: props.data.spanAnnotationId });
+      const edgeExists = reactFlowInstance.getEdges().some((edge) => edge.id === newEdge.id);
+      if (!edgeExists) {
+        reactFlowInstance.addEdges([newEdge]);
+      }
     }
   }, [props.data.spanAnnotationId, reactFlowInstance, memo.data]);
